[UI Fix] Moved variant_options_text to Admin.VariantView from Admin.OrderView
