+++ apps/ex_shop/web/views/admin/order_view.ex
@@ -1,6 +1,8 @@
 defmodule ExShop.Admin.OrderView do
   use ExShop.Web, :view
 
+  alias ExShop.Admin.VariantView
+
   def only_master_variant?(%ExShop.Product{variants: [_]}), do: true
   def only_master_variant?(%ExShop.Product{variants: [_|_]}), do: false
 
@@ -12,7 +14,7 @@ defmodule ExShop.Admin.OrderView do
       (variant) ->
         content_tag(:option, value: variant.id, disabled: out_of_stock?(variant)) do
         # TODO maybe autogenerate sku if one is not provided
-          (variant_options_text(variant))
+          (VariantView.variant_options_text(variant))
           <> if out_of_stock?(variant) do
             "(Out of stock)"
           else
@@ -28,23 +30,6 @@ defmodule ExShop.Admin.OrderView do
     ## Assuming everything pre-loaded
     variant = line_item.variant
     product = variant.product
-    product.name <> variant_options_text(variant)
-  end
-
-  # Calling from Admin.LineItemView
-  # so made Public
-  # TODO: Move to Common Module
-  def variant_options_text(variant) do
-    if variant.is_master do
-      ""
-    else
-      " (" <> (
-        Enum.map(variant.option_values,
-          fn(y) -> "#{y.option_type.presentation}:#{y.presentation}" end
-        )
-        |> Enum.join(", ")
-      )
-      <> ")"
-    end
+    product.name <> VariantView.variant_options_text(variant)
   end
 end
